+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.1] - 2020-01-22
+
+### Added
+
+- Add ARN of the IAM role used by the ALB Ingress controller as output variable.
+
+### Changed
+
+- Improve documentation of input variables.
+
 ## [3.0.0] - 2020-01-22
 
 ### Added